move DeepFormats to BTauReco (edit paths commit)
