+++ DataFormats/BTauReco/interface/DeepFlavourTagInfo.h
@@ -1,9 +1,9 @@
-#ifndef DataFormats_DeepFormats_DeepFlavourTagInfo_h
-#define DataFormats_DeepFormats_DeepFlavourTagInfo_h
+#ifndef DataFormats_BTauReco_DeepFlavourTagInfo_h
+#define DataFormats_BTauReco_DeepFlavourTagInfo_h
 
 #include "DataFormats/Common/interface/CMS_CLASS_VERSION.h"
 #include "DataFormats/BTauReco/interface/BaseTagInfo.h"
-#include "DataFormats/DeepFormats/interface/DeepFlavourFeatures.h"
+#include "DataFormats/BTauReco/interface/DeepFlavourFeatures.h"
 
 #include "DataFormats/PatCandidates/interface/Jet.h"
 
@@ -24,7 +24,7 @@ template<class Features> class FeaturesTagInfo : public BaseTagInfo {
 
     const Features & features() const { return features_; }
 
-    CMS_CLASS_VERSION(10)
+    CMS_CLASS_VERSION(3)
 
   private:
     Features features_;
@@ -37,4 +37,4 @@ DECLARE_EDM_REFS( DeepFlavourTagInfo )
 
 }
 
-#endif // DataFormats_DeepFormats_DeepFlavourTagInfo_h
+#endif // DataFormats_BTauReco_DeepFlavourTagInfo_h
