+++ DataFormats/BTauReco/interface/JetFeatures.h
@@ -1,7 +1,5 @@
-#ifndef DataFormats_DeepFormats_JetFeatures_h
-#define DataFormats_DeepFormats_JetFeatures_h
-
-#include "DataFormats/Common/interface/CMS_CLASS_VERSION.h"
+#ifndef DataFormats_BTauReco_JetFeatures_h
+#define DataFormats_BTauReco_JetFeatures_h
 
 namespace btagbtvdeep {
 
@@ -18,4 +16,4 @@ class JetFeatures {
 
 }
 
-#endif //DataFormats_DeepFormats_JetFeatures_h
+#endif //DataFormats_BTauReco_JetFeatures_h
