+++ DataFormats/BTauReco/interface/SecondaryVertexFeatures.h
@@ -1,7 +1,5 @@
-#ifndef DataFormats_DeepFormats_SecondaryVertexFeatures_h
-#define DataFormats_DeepFormats_SecondaryVertexFeatures_h
-
-#include "DataFormats/Common/interface/CMS_CLASS_VERSION.h"
+#ifndef DataFormats_BTauReco_SecondaryVertexFeatures_h
+#define DataFormats_BTauReco_SecondaryVertexFeatures_h
 
 namespace btagbtvdeep {
 
@@ -29,4 +27,4 @@ class SecondaryVertexFeatures {
 
 }
 
-#endif //DataFormats_DeepFormats_SecondaryVertexFeatures_h
+#endif //DataFormats_BTauReco_SecondaryVertexFeatures_h
