+++ RecoBTag/DeepFlavour/interface/NeutralCandidateConverter.h
@@ -2,7 +2,7 @@
 #define RecoSV_DeepFlavour_NeutralCandidateConverter_h
 
 #include "RecoBTag/DeepFlavour/interface/deep_helpers.h"
-#include "DataFormats/DeepFormats/interface/NeutralCandidateFeatures.h"
+#include "DataFormats/BTauReco/interface/NeutralCandidateFeatures.h"
 
 #include "DataFormats/PatCandidates/interface/PackedCandidate.h"
 #include "DataFormats/ParticleFlowCandidate/interface/PFCandidate.h"
